% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_fra_bailey_ware_abies.R
\name{scr_fra_bailey_ware_abies}
\alias{scr_fra_bailey_ware_abies}
\title{Sistema de crecimiento con la FRA de Bailey y Ware (1983)}
\usage{
scr_fra_bailey_ware_abies(
  E,
  Narb,
  IS,
  Thinning = FALSE,
  Ets = NULL,
  ABpts = NULL
)
}
\arguments{
\item{E, }{Vector de valores numérico. Al menos utilizar un vector con dos valores.}

\item{Narb, }{Valor numérico. Indica la densidad inicial del rodal - número de árboles por ha.}

\item{IS, }{Valor numérico. Indica el índice de sitio del rodal.}

\item{Thinning, }{Valor lógico. FALSE no simula el crecimiento con efecto de aclareo. TRUE incluye el efecto de aclareo}

\item{Ets, }{Valor o vector de valores numérico. Indica la edad del rodal cuando se aplica el aclareo.}

\item{ABpts, }{Valor o vector de valores numérico. Indica la intensidad de aclareo respecto al AB, por ejemplo 0.3, equivale al 30 \%.}
}
\value{
Regresa un objeto tibble.
}
\description{
Simulador de crecimiento de atributos de rodal como altura dominante, mortalidad y área basal
sin o con efecto de aclareo.
}
\examples{
# scr_fra_bailey_ware_abies(E = 1:5, Narb = 2000, IS = 28)


}
